Fix C++20 warnings about operations on enums of different types.
C++20 deprecates support for bitwise operations where the two arguments
have different enum types. Compilers generate warnings about this.

In Headers.cxx, commaHandling is used in bitwise operations with
ParserCategory::CommasAllowedOutputMulti. Since commaHandling is defined
as an unnamed enum in each class derived from ParserCategory, each
constant constitutes a distinct enum type. To avoid the warnings,
name the enum in ParserCategory and use that name to define commaHandling
constants.

In DumFeatureChain.cxx and DialogUsageManager.cxx, DumFeature::ProcessingResult
bit masks are tested using DumFeature::ProcessingResultMask bits, which
are two different enums. To avoid the warnings, cast the enums to
unsigned int (since these are bit masks).

+++ resip/dum/DialogUsageManager.cxx
@@ -1067,13 +1067,13 @@ void DialogUsageManager::outgoingProcess(std::unique_ptr<Message> message)
       
       DumFeatureChain::ProcessingResult res = it->second->process(message.get());
       
-      if (res & DumFeatureChain::ChainDoneBit)
+      if (static_cast<unsigned int>(res) & DumFeatureChain::ChainDoneBit)
       {
          delete it->second;
          mOutgoingFeatureChainMap.erase(it);
       }
 
-      if (res & DumFeatureChain::EventTakenBit)
+      if (static_cast<unsigned int>(res) & DumFeatureChain::EventTakenBit)
       {
          message.release();
          return;
@@ -1608,14 +1608,14 @@ DialogUsageManager::incomingProcess(std::unique_ptr<Message> msg)
       
       DumFeatureChain::ProcessingResult res = it->second->process(msg.get());
       
-      if (res & DumFeatureChain::ChainDoneBit)
+      if (static_cast<unsigned int>(res) & DumFeatureChain::ChainDoneBit)
       {
          delete it->second;
          mIncomingFeatureChainMap.erase(it);
          //DebugLog(<< "feature chain deleted" << endl);
       }
  
-      if (res & DumFeatureChain::EventTakenBit)
+      if (static_cast<unsigned int>(res) & DumFeatureChain::EventTakenBit)
       {
          msg.release();
          //DebugLog(<< "event taken");
